requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
gunicorn==23.0.0

pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

orjson==3.11.5
httpx==0.28.1
tenacity==9.1.2
structlog==25.5.0

# Data layer
sqlalchemy[asyncio]==2.0.45
asyncpg==0.31.0
alembic==1.17.2

# Redis (async: `from redis import asyncio as redis_async`)
redis==7.0.1

# ClickHouse
clickhouse-connect==0.10.0
